--- feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/database/converter/AlbumImageEntityTypeConverter.kt
@@ -1,21 +1,0 @@
-package com.igorwojda.showcase.feature.album.data.database.converter
-
-import androidx.room.TypeConverter
-import com.igorwojda.showcase.feature.album.data.database.model.AlbumImageEntity
-import com.squareup.moshi.JsonAdapter
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.Types
-
-class AlbumImageEntityTypeConverter {
-
-    private val type = Types.newParameterizedType(List::class.java, AlbumImageEntity::class.java)
-    private val adapter: JsonAdapter<List<AlbumImageEntity>> = Moshi.Builder().build().adapter(type)
-
-    @TypeConverter
-    fun stringToList(data: String?) =
-        data?.let { adapter.fromJson(it) } ?: listOf()
-
-    @TypeConverter
-    fun listToString(someObjects: List<AlbumImageEntity>): String =
-        adapter.toJson(someObjects)
-}
